requirements.txt
Flask==3.1.0
openai==1.65.2
PyPDF2==3.0.1
gTTS==2.2.4
langdetect==1.0.9
gunicorn==20.1.0
pyautogen>=0.1.0
